# feat: eliminate array operations for 25% performance boost