--- a/package.json
+++ b/package.json
@@ -38,7 +38,7 @@
   "prettier": "@shelf/prettier-config",
   "dependencies": {},
   "devDependencies": {
-    "@shelf/eslint-config": "4.4.0",
+    "@shelf/eslint-config": "5.2.3",
     "@shelf/prettier-config": "1.0.0",
     "@shelf/tsconfig": "0.1.0",
     "@swc/core": "1.10.18",
@@ -47,7 +47,7 @@
     "@types/node": "20.17.10",
     "@types/uslug": "1.0.4",
     "benny": "3.7.1",
-    "eslint": "9.30.0",
+    "eslint": "9.31.0",
     "fast-lorem-ipsum": "1.2.0",
     "husky": "9.1.7",
     "jest": "29.7.0",
@@ -55,7 +55,7 @@
     "lint-staged": "15.5.2",
     "prettier": "3.6.2",
     "ts-jest-resolver": "2.0.1",
-    "tsx": "4.19.3",
+    "tsx": "4.20.3",
     "typescript": "5.8.3",
     "uslug": "1.0.4"
   },