## Summary
Replace array.push() + join() with direct string concatenation in slug generation for significant performance improvements.

## Performance Impact
- **10 words**: 287,961 → 359,774 ops/s (+24.9%)
- **100 words**: 25,391 → 31,670 ops/s (+24.7%)  
- **1000 words**: 2,405 → 3,124 ops/s (+29.9%)
- **Overall**: 17x → 20x faster than uslug

## Changes
- Replaced `rawSlug.push()` operations with `rawSlug += char`
- Eliminated final `join('')` call in favor of direct string building
- Reduces memory allocations and garbage collection pressure
- Maintains identical functionality and test compatibility

## Test Plan
- [x] All existing tests pass
- [x] Benchmark results show consistent 25-30% improvement
- [x] No breaking changes to API or behavior

🤖 Generated with [Claude Code](https://claude.ai/code)